adds basic jsdoc generation
I've added the basics of jsdocs and a starting point for adding jsdocs
to the components and modules in xelah-core, etc.

Styleguidest provides good documentation for components but it doesn't
provide nearly the tooling that jsdocs provides for encouraging good
code organization and documentation. For example styleguidest doesn't
provide documenting module exports.

EditableBlock is the simplest component we have so I've started there.
I wouldn't add CI for publishing documentation just yet - maybe by the
end of collab week?