adds jsdoc for params for EditableBlock
I'd like to investigating moving away from styleguidest and this
is a first step. Since react components are functions that take
their data and return JSX we can document as such using jsdoc and
leverage the other features jsdoc provides (links etc.).

I realize that a move away from styleguidest may leave us unable to
have "live" previews of components, however, from my experience in
the C# and Haskell communities this typically isn't a problem for most
developers - that is most developers get a long quite well without it.
Furthermore it's less to maintain. I realize this is very oppinionated
of me and so this may not be the best direction to go in.